issue #3: JMS logos + final tweaks
