+++ moo-client-jms/pom.xml
@@ -7,6 +7,20 @@
     <version>1.3.0-SNAPSHOT</version>
   </parent>
   <artifactId>moo-client-jms</artifactId>
+  
+  <build>
+    <resources>
+      <resource>
+        <directory>src/main/resources</directory>
+        <filtering>false</filtering>
+      </resource>
+      <resource>
+        <directory>src/main/resources-filtered</directory>
+        <filtering>true</filtering>
+      </resource>
+    </resources>
+  </build>
+  
   <dependencies>
     <dependency>
       <groupId>pl.zimowski</groupId>
