+++ moo-client-jms/src/main/resources-filtered/logo
@@ -1,0 +1,6 @@
+   (___)
+   (o o)_____/
+    @@ `     \        moo!    v:${project.version}
+     \ ____, /        JMS CLIENT
+     //    //  jgs
+    ^^    ^^
