+++ moo-jms/pom.xml
@@ -13,6 +13,10 @@
       <artifactId>cdi-api</artifactId>
     </dependency>
     <dependency>
+      <groupId>org.springframework</groupId>
+      <artifactId>spring-context</artifactId>
+    </dependency>
+    <dependency>
       <groupId>javax.jms</groupId>
       <artifactId>javax.jms-api</artifactId>
     </dependency>
