+++ moo-server-jms/src/main/java/pl/zimowski/moo/server/jms/JmsChatService.java
@@ -2,11 +2,15 @@ package pl.zimowski.moo.server.jms;
 
 import java.util.Scanner;
 
+import javax.annotation.PreDestroy;
 import javax.inject.Inject;
 
 import org.slf4j.Logger;
 import org.springframework.stereotype.Component;
 
+import pl.zimowski.moo.api.ApiUtils;
+import pl.zimowski.moo.api.ServerAction;
+import pl.zimowski.moo.api.ServerEvent;
 import pl.zimowski.moo.jms.JmsGateway;
 import pl.zimowski.moo.server.commons.ChatService;
 
@@ -25,21 +29,25 @@ public class JmsChatService implements ChatService {
 
     private boolean running;
 
+    @Inject
+    private ServerEventPublisher publisher;
+
 
     @Override
     public void start() {
 
-        // model after jndi.properties example from https://dzone.com/articles/jms-activemq
+        log.info("\n{}", ApiUtils.fetchResource("/logo"));
 
         try(Scanner scanner = new Scanner(System.in)) {
 
             running = true;
 
-            log.info("server running (type moo:exit to stop)");
+            log.info("server running (moo:exit or CTL-C to stop)");
 
             while(running) {
                 String input = scanner.nextLine();
                 if(input.equals("moo:exit")) {
+                    notifyClients();
                     jms.stop();
                     running = false;
                 }
@@ -48,6 +56,13 @@ public class JmsChatService implements ChatService {
         catch(Exception e) {
             running = false;
             log.error("jms service init error", e);
+        }
+    }
+
+    @PreDestroy
+    void notifyClients() {
+        if(running) {
+            publisher.publish(new ServerEvent(ServerAction.ServerExit));
         }
     }
