+++ moo-server-jms/src/main/resources-filtered/logo
@@ -1,0 +1,5 @@
+    ()___()
+    < @ @ >
+     |   |   moo ...   v:${project.version}
+     {o_o}   JMS SERVER
+      (|)
