+++ moo-ui-shell/moo-ui-shell-writer/src/main/java/pl/zimowski/moo/ui/shell/writer/App.java
@@ -69,6 +69,8 @@ public class App implements ApplicationRunner {
     @Override
     public void run(ApplicationArguments args) throws Exception {
 
+        ClientReporter.LOG.info("\n{}", ApiUtils.fetchResource("/logo"));
+
         try (Scanner scanner = new Scanner(System.in)) {
 
             if(!clientHandler.connect(clientReporter)) {
